DzukkuBot/requirements.txt
# Core
python-dotenv==1.0.1
pandas==2.2.2

# Gemini AI
google-generativeai>=0.8.0

# Telegram
python-telegram-bot==21.0.1

# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Google Sheets sync
gspread==6.2.1
google-auth==2.35.0

# Utilities
requests==2.32.3

# Excel loader (Master_Menu sheet in data/Project_Dzukku.xlsx)
openpyxl>=3.1.0

# ── PostgreSQL + ORM + Migrations (vNext) ──────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0

# ── Auth (JWT) ──────────────────────────────────────────────────────────────
pyjwt>=2.8.0

# ── Payments (Razorpay) ────────────────────────────────────────────────────
httpx>=0.27.0

# ── Agent policies ──────────────────────────────────────────────────────────
pytz>=2024.1

# ── Object storage (optional — install only if using S3/GCS/Azure) ──────────
# aioboto3>=13.0.0        # for S3
# google-cloud-storage>=2.0.0  # for GCS
# azure-storage-blob>=12.0.0   # for Azure

# ── LangGraph + MCP (Zomato / Swiggy ordering via Model Context Protocol) ──
# stdio MCP client + LangGraph ReAct agent that binds MCP tools to Gemini.
# mcp-remote (Node/npx) is spawned as a subprocess to bridge HTTP MCP → stdio
# and handle OAuth on its own loopback callback (no public callback needed).
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
mcp>=1.0.0
